c2

+++ BackEnd/server.js
@@ -17,7 +17,7 @@ app.use(bp.json());
 app.use(bp.urlencoded({ extended: true }));
 
 
-app.get("/data/:id/:myname/:age/:del/:ins/:oiname/:oiage/:upd", function (req, res) {
+app.get("/data/:id/:myname/:age/:del/:ins/:oiname/:oiage/:upd/:srch", function (req, res) {
  
   var fnamee = req.params.myname;
   var fnameesq = mysql.escape(fnamee);
@@ -32,10 +32,12 @@ app.get("/data/:id/:myname/:age/:del/:ins/:oiname/:oiage/:upd", function (req, r
   var del = req.params.del;
   var age = req.params.age;
   var upd = req.params.upd;
+  var srch = req.params.srch;
+  
 
   
 //DELETE
-if(id!=0 && del!=0 && upd==0 && ins==0){
+if(id!=0 && del!=0 && upd==0 && ins==0 && srch==0){
  
   var sql1=`delete from labtbl6 where id=${id}`;}
   else{
@@ -49,7 +51,7 @@ if(id!=0 && del!=0 && upd==0 && ins==0){
   
 
   //INSERT
-  if (oiname!= 0 && oiage!= 0 && oiname!= '' && oiage!= '' && del == 0 && ins==1 && upd==0 )
+  if (oiname!= 0 && oiage!= 0 && oiname!= '' && oiage!= '' && oiname!=' ' && oiage!=' ' && del == 0 && ins==1 && upd==0 && srch==0 )
     var sql0 = `insert into labtbl6 (fname,age)values(${uinameesq},${oiage})`;
   else var sql0 = "SELECT * FROM labtbl6";
 
@@ -58,7 +60,7 @@ if(id!=0 && del!=0 && upd==0 && ins==0){
   });
   
  //UPDATE
-if(id!=0 && fnameesq!=0 && fnameesq!='' && age!=0 && del==0 && upd==1 && ins==0)
+if(id!=0 && fnameesq!=0 && fnameesq!='' && fnameesq!='undefined' && age!=0 && age!='undefined' && del==0 && upd==1 && ins==0 && srch==0)
 var sql10=`update labtbl6 set fname=${fnameesq}, age=${age} where id=${id}`;
 else
 var sql10=`SELECT * FROM labtbl6`;
@@ -67,15 +69,35 @@ con.query(sql10, function (err, data, fields) {
   if (err) throw err;
 });
 
+//SEARCH
+if(oiname!=0 && oiage!='' && oiname!=' ' && oiage!=' ' && (oiname!='undefined' || oiage!='undefined') && del==0 && upd==0 && ins==0 && srch==1)
+{
+  if(oiname!='undefined' && oiage=='undefined')
+var sqls=`select * from labtbl6 where fname=${uinameesq}`;
+if(oiname!='undefined' && oiage!='undefined')
+var sqls=`select * from labtbl6 where fname=${uinameesq} OR age=${oiage}`;
+if(oiname=='undefined' && oiage!='undefined')
+var sqls=`select * from labtbl6 where age=${oiage}`;
+
+
+}
+else{
+var sqls=`SELECT * FROM labtbl6`;}
+
+con.query(sqls, function (err, data, fields) {
+  if (err) throw err;
+  res.send({ userData: data });
+});
 
 
 
-  var sql=`SELECT * FROM labtbl6`;
 
-  con.query("SELECT * FROM labtbl6", function (err, data, fields) {
-    if (err) throw err;
-    res.send({ userData: data });
-  });
+  // var sql=`SELECT * FROM labtbl6`;
+
+  // con.query("select * from labtbl6", function (err, data, fields) {
+  //   if (err) throw err;
+  //   res.send({ userData: data });
+  // });
 });
